[FIX/REFACTOR] updateCamera now updates keyframe, parseSetCameraType now a generic util
